+++ CRM/Models/Student.cs
@@ -8,7 +8,7 @@ namespace CRM.Models
     public class Student
     {
         public int Id { get; set; }
-        public int? AdmissionFormNo { get; set; }
+        public long? AdmissionFormNo { get; set; }
         public string? Year { get; set; }
         public string? Session { get; set; }
         public string? EnRollNo { get; set; }
