+++ CRM/ModelsForView/StudentRegDownload.cs
@@ -9,7 +9,7 @@ namespace CRM.ModelsForView
     {
         public int Id { get; set; }
         public int? RegNo { get; set; }
-        public int? FormNo { get; set; }
+        public long? FormNo { get; set; }
         public int? SchoNo { get; set; }
         public string? Session { get; set; }
         public string? Year { get; set; }
